Remove old CollectionList component, rename new component
